LayoutTests/fast/dom/repeating-timer-element-overflow-throttling-expected.txt

Tests that a repeating timer changing the style of an element with 0 height and visible overflowing content does not get throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testElement.offsetHeight is 0
The timer should initially not be throttled.
PASS internals.isTimerThrottled(timeoutId) is false
5th iteration, timer should not have been throttled.
PASS wasThrottled is false
6th iteration, timer should still be unthrottled.
PASS internals.isTimerThrottled(timeoutId) is false
PASS successfullyParsed is true

TEST COMPLETE
test1
test2
